Add utiltiy methods to GraphWidget
similar API to VisualizationGraph going forward. this will avoid rerendering for every change to take effect
